requirements.txt
streamlit>=1.20.0
python-dotenv>=1.0.0
diffusers>=0.20.0
transformers>=4.40.0
accelerate>=0.18.0
safetensors>=0.3.0
torch>=2.0.0
Pillow>=9.5.0
openai>=1.11.0
numpy>=1.24.0
requests>=2.31.0
